Moved shared scanning code to ScanUtil
* Deprecated all classes in fluo-cluster module
