+++ modules/cluster/src/main/java/org/apache/fluo/cluster/runnable/WorkerRunnable.java
@@ -33,6 +33,7 @@ import org.slf4j.LoggerFactory;
 /**
  * Run method of Fluo worker that is called within a Twill/YARN application
  */
+@Deprecated
 public class WorkerRunnable extends AbstractTwillRunnable {
 
   private static final Logger log = LoggerFactory.getLogger(WorkerRunnable.class);
