+++ modules/cluster/src/main/java/org/apache/fluo/cluster/runner/AppRunner.java
@@ -17,45 +17,32 @@ package org.apache.fluo.cluster.runner;
 
 import java.lang.reflect.Method;
 import java.util.Arrays;
-import java.util.Collection;
-import java.util.HashSet;
-
+import java.util.Collections;
+import java.util.List;
 import javax.inject.Provider;
 
 import com.beust.jcommander.JCommander;
+import com.beust.jcommander.Parameter;
 import com.beust.jcommander.ParameterException;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.Iterables;
 import com.google.inject.AbstractModule;
 import com.google.inject.Guice;
-import org.apache.accumulo.core.client.Connector;
 import org.apache.accumulo.core.client.Scanner;
 import org.apache.accumulo.core.client.TableNotFoundException;
-import org.apache.accumulo.core.security.Authorizations;
-import org.apache.fluo.accumulo.format.FluoFormatter;
-import org.apache.fluo.api.client.FluoClient;
-import org.apache.fluo.api.client.FluoFactory;
-import org.apache.fluo.api.client.Snapshot;
-import org.apache.fluo.api.client.scanner.CellScanner;
 import org.apache.fluo.api.config.FluoConfiguration;
-import org.apache.fluo.api.data.Bytes;
-import org.apache.fluo.api.data.Column;
-import org.apache.fluo.api.data.RowColumnValue;
-import org.apache.fluo.api.data.Span;
 import org.apache.fluo.api.exceptions.FluoException;
 import org.apache.fluo.cluster.util.FluoYarnConfig;
 import org.apache.fluo.core.impl.Environment;
 import org.apache.fluo.core.impl.Notification;
-import org.apache.fluo.core.util.AccumuloUtil;
-import org.apache.fluo.core.util.ByteUtil;
-import org.apache.fluo.core.util.Hex;
-import org.apache.fluo.core.util.SpanUtil;
+import org.apache.fluo.core.util.ScanUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
  * Base class for running a Fluo application
  */
+@Deprecated
 public abstract class AppRunner {
 
   private static final Logger log = LoggerFactory.getLogger(AppRunner.class);
@@ -68,62 +55,7 @@ public abstract class AppRunner {
     this.scriptName = scriptName;
   }
 
-  public static Span getSpan(ScanOptions options) {
-    Span span = new Span();
-    if ((options.getExactRow() != null)
-        && ((options.getStartRow() != null) || (options.getEndRow() != null) || (options
-            .getRowPrefix() != null))) {
-      throw new IllegalArgumentException(
-          "You cannot specify an exact row with a start/end row or row prefix!");
-    }
-
-    if ((options.getRowPrefix() != null)
-        && ((options.getStartRow() != null) || (options.getEndRow() != null) || (options
-            .getExactRow() != null))) {
-      throw new IllegalArgumentException(
-          "You cannot specify an prefix row with a start/end row or exact row!");
-    }
-
-    // configure span of scanner
-    if (options.getExactRow() != null) {
-      span = Span.exact(options.getExactRow());
-    } else if (options.getRowPrefix() != null) {
-      span = Span.prefix(options.getRowPrefix());
-    } else {
-      if ((options.getStartRow() != null) && (options.getEndRow() != null)) {
-        span = new Span(options.getStartRow(), true, options.getEndRow(), true);
-      } else if (options.getStartRow() != null) {
-        span = new Span(Bytes.of(options.getStartRow()), true, Bytes.EMPTY, true);
-      } else if (options.getEndRow() != null) {
-        span = new Span(Bytes.EMPTY, true, Bytes.of(options.getEndRow()), true);
-      }
-    }
-
-    return span;
-  }
-
-  public static Collection<Column> getColumns(ScanOptions options) {
-    Collection<Column> columns = new HashSet<>();
-
-    // configure columns of scanner
-    for (String column : options.getColumns()) {
-      String[] colFields = column.split(":");
-      if (colFields.length == 1) {
-        columns.add(new Column(colFields[0]));
-      } else if (colFields.length == 2) {
-        columns.add(new Column(colFields[0], colFields[1]));
-      } else {
-        throw new IllegalArgumentException("Failed to scan!  Column '" + column
-            + "' has too many fields (indicated by ':')");
-      }
-    }
-
-    return columns;
-  }
-
-
-
-  public long scan(FluoConfiguration config, String[] args) {
+  public void scan(FluoConfiguration config, String[] args) {
     ScanOptions options = new ScanOptions();
     JCommander jcommand = new JCommander(options);
     jcommand.setProgramName(scriptName + " scan <app>");
@@ -141,106 +73,10 @@ public abstract class AppRunner {
     }
 
     if (options.scanAccumuloTable) {
-      return scanAccumulo(options, config);
+      ScanUtil.scanAccumulo(options.getScanOpts(), config);
     } else {
-      return scanFluo(options, config);
+      ScanUtil.scanFluo(options.getScanOpts(), config);
     }
-  }
-
-  private long scanFluo(ScanOptions options, FluoConfiguration sConfig) {
-    System.out.println("Scanning snapshot of data in Fluo '" + sConfig.getApplicationName()
-        + "' application.");
-
-    long entriesFound = 0;
-    try (FluoClient client = FluoFactory.newClient(sConfig)) {
-      try (Snapshot s = client.newSnapshot()) {
-
-        Span span = null;
-        Collection<Column> columns = null;
-        try {
-          span = getSpan(options);
-          columns = getColumns(options);
-        } catch (IllegalArgumentException e) {
-          System.err.println(e.getMessage());
-          System.exit(-1);
-        }
-
-        CellScanner cellScanner = s.scanner().over(span).fetch(columns).build();
-
-        StringBuilder sb = new StringBuilder();
-        for (RowColumnValue rcv : cellScanner) {
-          if (options.hexEncNonAscii) {
-            sb.setLength(0);
-            Hex.encNonAscii(sb, rcv.getRow());
-            sb.append(" ");
-            Hex.encNonAscii(sb, rcv.getColumn(), " ");
-            sb.append("\t");
-            Hex.encNonAscii(sb, rcv.getValue());
-            System.out.println(sb.toString());
-          } else {
-            sb.setLength(0);
-            sb.append(rcv.getsRow());
-            sb.append(" ");
-            sb.append(rcv.getColumn());
-            sb.append("\t");
-            sb.append(rcv.getsValue());
-            System.out.println(sb.toString());
-          }
-          entriesFound++;
-          if (System.out.checkError()) {
-            break;
-          }
-        }
-
-        if (entriesFound == 0) {
-          System.out.println("\nNo data found\n");
-        }
-      } catch (FluoException e) {
-        System.out.println("Scan failed - " + e.getMessage());
-      }
-    }
-    return entriesFound;
-  }
-
-  private long scanAccumulo(ScanOptions options, FluoConfiguration sConfig) {
-    System.out.println("Scanning data in Accumulo directly for '" + sConfig.getApplicationName()
-        + "' application.");
-
-    Connector conn = AccumuloUtil.getConnector(sConfig);
-
-    Span span = null;
-    Collection<Column> columns = null;
-    try {
-      span = getSpan(options);
-      columns = getColumns(options);
-    } catch (IllegalArgumentException e) {
-      System.err.println(e.getMessage());
-      System.exit(-1);
-    }
-
-    long entriesFound = 0;
-
-    try {
-      Scanner scanner = conn.createScanner(sConfig.getAccumuloTable(), Authorizations.EMPTY);
-      scanner.setRange(SpanUtil.toRange(span));
-      for (Column col : columns) {
-        if (col.isQualifierSet()) {
-          scanner
-              .fetchColumn(ByteUtil.toText(col.getFamily()), ByteUtil.toText(col.getQualifier()));
-        } else {
-          scanner.fetchColumnFamily(ByteUtil.toText(col.getFamily()));
-        }
-      }
-
-      for (String entry : Iterables.transform(scanner, FluoFormatter::toString)) {
-        System.out.println(entry);
-      }
-    } catch (Exception e) {
-      System.out.println("Scan failed - " + e.getMessage());
-      entriesFound++;
-    }
-
-    return entriesFound;
   }
 
   private long calculateSleep(long notifyCount, long numWorkers) {
@@ -320,7 +156,6 @@ public abstract class AppRunner {
         }
       });
     }
-
   }
 
   public void exec(FluoConfiguration fluoConfig, String[] args) throws Exception {
@@ -335,5 +170,66 @@ public abstract class AppRunner {
 
     Method method = clazz.getMethod("main", String[].class);
     method.invoke(null, (Object) Arrays.copyOfRange(args, 1, args.length));
+  }
+
+  public static class ScanOptions {
+
+    @Parameter(names = "-s", description = "Start row (inclusive) of scan")
+    private String startRow;
+
+    @Parameter(names = "-e", description = "End row (inclusive) of scan")
+    private String endRow;
+
+    @Parameter(names = "-c", description = "Columns of scan in comma separated format: "
+        + "<<columnfamily>[:<columnqualifier>]{,<columnfamily>[:<columnqualifier>]}> ")
+    private List<String> columns;
+
+    @Parameter(names = "-r", description = "Exact row to scan")
+    private String exactRow;
+
+    @Parameter(names = "-p", description = "Row prefix to scan")
+    private String rowPrefix;
+
+    @Parameter(names = {"-h", "-help", "--help"}, help = true, description = "Prints help")
+    public boolean help;
+
+    @Parameter(names = {"-esc", "--escape-non-ascii"}, help = true,
+        description = "Hex encode non ascii bytes", arity = 1)
+    public boolean hexEncNonAscii = true;
+
+    @Parameter(
+        names = "--raw",
+        help = true,
+        description = "Show underlying key/values stored in Accumulo. Interprets the data using Fluo "
+            + "internal schema, making it easier to comprehend.")
+    public boolean scanAccumuloTable = false;
+
+    public String getStartRow() {
+      return startRow;
+    }
+
+    public String getEndRow() {
+      return endRow;
+    }
+
+    public String getExactRow() {
+      return exactRow;
+    }
+
+    public String getRowPrefix() {
+      return rowPrefix;
+    }
+
+    public List<String> getColumns() {
+      if (columns == null) {
+        return Collections.emptyList();
+      }
+      return columns;
+    }
+
+    public ScanUtil.ScanOpts getScanOpts() {
+      return new ScanUtil.ScanOpts(startRow, endRow, columns, exactRow, rowPrefix, help,
+          hexEncNonAscii, scanAccumuloTable);
+    }
   }
 }
