+++ modules/cluster/src/main/java/org/apache/fluo/cluster/util/FluoInstall.java
@@ -20,6 +20,7 @@ import java.io.File;
 import org.apache.fluo.api.config.FluoConfiguration;
 import org.apache.fluo.api.exceptions.FluoException;
 
+@Deprecated
 public class FluoInstall {
 
   private String fluoHomeDir;
