+++ modules/cluster/src/main/java/org/apache/fluo/cluster/yarn/TwillUtil.java
@@ -23,6 +23,7 @@ import org.apache.twill.api.TwillRunResources;
 /**
  * Twill Utility classes
  */
+@Deprecated
 public class TwillUtil {
 
   private TwillUtil() {}
